Move structs into WavedataStructs and move read methods to WaveReader.h
