+++ WaveWriter.h
@@ -1,0 +1,2 @@
+
+#include "WaveDataStructs.h"
